+++ app/components/DefaultCollectionInputSelect.tsx
@@ -5,20 +5,14 @@ import CollectionIcon from "~/components/CollectionIcon";
 import Flex from "~/components/Flex";
 import InputSelect from "~/components/InputSelect";
 import { IconWrapper } from "~/components/Sidebar/components/SidebarLink";
+import useCurrentTeam from "~/hooks/useCurrentTeam";
 import useStores from "~/hooks/useStores";
 import useToasts from "~/hooks/useToasts";
 
-type DefaultCollectionInputSelectProps = {
-  onSelectCollection: (value: string) => void;
-  defaultCollectionId?: string | null;
-};
-
-const DefaultCollectionInputSelect = ({
-  onSelectCollection,
-  defaultCollectionId,
-}: DefaultCollectionInputSelectProps) => {
+const DefaultCollectionInputSelect = () => {
   const { t } = useTranslation();
-  const { collections } = useStores();
+  const { collections, auth } = useStores();
+  const team = useCurrentTeam();
   const [fetching, setFetching] = useState(false);
   const [fetchError, setFetchError] = useState();
   const { showToast } = useToasts();
@@ -47,45 +41,68 @@ const DefaultCollectionInputSelect = ({
     load();
   }, [showToast, fetchError, t, fetching, collections]);
 
-  const options = collections.publicCollections.reduce(
-    (acc, collection) => [
-      ...acc,
-      {
-        label: (
-          <Flex align="center">
-            <IconWrapper>
-              <CollectionIcon collection={collection} />
-            </IconWrapper>
-            {collection.name}
-          </Flex>
-        ),
-        value: collection.id,
-      },
-    ],
-    [
-      {
-        label: (
-          <Flex align="center">
-            <IconWrapper>
-              <HomeIcon color="currentColor" />
-            </IconWrapper>
-            {t("Home")}
-          </Flex>
-        ),
-        value: "home",
-      },
-    ]
+  const onSelectCollection = React.useCallback(
+    async (value: string) => {
+      const defaultCollectionId = value === "home" ? null : value;
+      try {
+        await auth.updateTeam({
+          defaultCollectionId,
+        });
+        showToast(t("Settings saved"), {
+          type: "success",
+        });
+      } catch (err) {
+        showToast(err.message, {
+          type: "error",
+        });
+      }
+    },
+    [auth, showToast, t]
+  );
+
+  const options = React.useMemo(
+    () =>
+      collections.publicCollections.reduce(
+        (acc, collection) => [
+          ...acc,
+          {
+            label: (
+              <Flex align="center">
+                <IconWrapper>
+                  <CollectionIcon collection={collection} />
+                </IconWrapper>
+                {collection.name}
+              </Flex>
+            ),
+            value: collection.id,
+          },
+        ],
+        [
+          {
+            label: (
+              <Flex align="center">
+                <IconWrapper>
+                  <HomeIcon color="currentColor" />
+                </IconWrapper>
+                {t("Home")}
+              </Flex>
+            ),
+            value: "home",
+          },
+        ]
+      ),
+    [collections.publicCollections, t]
   );
 
   if (fetching) return null;
 
   return (
     <InputSelect
-      value={defaultCollectionId ?? "home"}
+      value={team.defaultCollectionId ?? "home"}
       label={t("Collection")}
       options={options}
       onChange={onSelectCollection}
-      ariaLabel={t("Perferred collection")}
+      ariaLabel={t("Default collection")}
       note={t(
         "We will redirect users to the selected collection when they sign in."
       )}
