+++ shared/i18n/locales/en_US/translation.json
@@ -63,7 +63,8 @@
   "Edits you make will sync once you’re online": "Edits you make will sync once you’re online",
   "Submenu": "Submenu",
   "Collections could not be loaded, please reload the app": "Collections could not be loaded, please reload the app",
-  "Perferred collection": "Perferred collection",
+  "Settings saved": "Settings saved",
+  "Default collection": "Default collection",
   "We will redirect users to the selected collection when they sign in.": "We will redirect users to the selected collection when they sign in.",
   "Deleted Collection": "Deleted Collection",
   "Unpin": "Unpin",
@@ -527,7 +528,6 @@
   "by {{ name }}": "by {{ name }}",
   "Last accessed": "Last accessed",
   "Add to Slack": "Add to Slack",
-  "Settings saved": "Settings saved",
   "document published": "document published",
   "document updated": "document updated",
   "Posting to the <em>{{ channelName }}</em> channel on": "Posting to the <em>{{ channelName }}</em> channel on",
